# openapi.yaml
openapi: 3.1.0
info:
  title: Code All Remote API
  version: 1.0.0
  description: Versioned control-plane API for self-hosted coding-agent sessions.
servers:
  - url: https://car.example.invalid/api/v1
paths:
...
  /sessions:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Authorized session summaries.
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/SessionSnapshot' }
    post:
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionSnapshot' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CAR-access-token
  responses:
...
    Conflict:
      description: State or idempotency conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [workspace_id, adapter_id]
      properties:
        workspace_id: { type: string }
        adapter_id: { type: string }
        title: { type: string, maxLength: 200 }
    SessionSnapshot:
      type: object
      additionalProperties: false
      required: [id, workspace_id, adapter_id, state, last_sequence]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        adapter_id: { type: string }
        state: { type: string }
        last_sequence: { type: integer, minimum: 0 }
        pending_approval_id: { type: [string, 'null'] }
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
        request_id: { type: string }